fix: vtap exceptions type

+++ server/controller/db/metadb/migrator/schema/rawsql/mysql/init.sql
@@ -1002,7 +1002,7 @@ CREATE TABLE IF NOT EXISTS vtap (
     synced_analyzer_at      DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP,
     created_at              DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP,
     boot_time               INTEGER DEFAULT 0,
-    exceptions              INTEGER UNSIGNED DEFAULT 0,
+    exceptions              BIGINT UNSIGNED DEFAULT 0,
     vtap_lcuuid             CHAR(64) DEFAULT NULL,
     vtap_group_lcuuid       CHAR(64) DEFAULT NULL,
     cpu_num                 INTEGER DEFAULT 0 COMMENT 'logical number of cpu',
